chore: rename status_checker.py to status.py and use 'incomplete' terminology
- Rename ami/jobs/status_checker.py → ami/jobs/status.py
- Replace 'unfinished jobs' with 'incomplete jobs' throughout
- Update migration file names and task names
- Update test class names
- Fix SourceImageCollectionAdmin missing search_fields
- Add git commit guidelines to CLAUDE.md

This makes terminology more accurate and sets up clean history for
subsequent feature additions.
